APPS: break out app_malloc() and app_bail_out() from apps/lib/apps.c
This is to fulfill a need, apps/lib/vms_decc_argv.c uses app_malloc(),
but the program using it doesn't necessarily want the rest of
apps/lib/apps.c.
