+++ apps/lib/app_mem.c
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2015-2021 The OpenSSL Project Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License 2.0 (the "License").  You may not use
+ * this file except in compliance with the License.  You can obtain a copy
+ * in the file LICENSE in the source distribution or at
+ * https://www.openssl.org/source/license.html
+ */
+
+#include <openssl/crypto.h>      /* OPENSSL_malloc() */
+#include "apps.h"
+
+void *app_malloc(size_t sz, const char *what)
+{
+    void *vp = OPENSSL_malloc(sz);
+
+    if (vp == NULL)
+        app_bail_out("%s: Could not allocate %zu bytes for %s\n",
+                     opt_getprog(), sz, what);
+    return vp;
+}
+
